Address comments, rearrange endpoints
