+++ packages/wallet/backend/src/card/types.ts
@@ -102,38 +102,3 @@ export interface ICardLockRequest {
 export interface ICardUnlockRequest {
   note: string
 }
-
-// Response for fetching card transactions
-export interface ITransaction {
-  id: number
-  transactionId: string
-  ghResponseCode: string
-  cardScheme: number
-  type: number
-  createdAt: string
-  txStatus: string
-  vaultId: number
-  cardId: number
-  refTransactionId: string
-  responseCode: string | null
-  transactionAmount: string
-  transactionCurrency: string
-  billingAmount: string
-  billingCurrency: string
-  terminalId: string | null
-  wallet: number
-  transactionDateTime: string
-  processDateTime: string | null
-}
-
-export interface IPagination {
-  pageNumber: number
-  pageSize: number
-  totalPages: number
-  totalRecords: number
-}
-
-export interface IGetTransactionsResponse {
-  data: ITransaction[]
-  pagination: IPagination
-}
