+++ packages/wallet/shared/src/types/card.ts
@@ -5,3 +5,38 @@ export type LockReasonCode =
   | 'IssuerRequestGeneral'
   | 'IssuerRequestFraud'
   | 'IssuerRequestLegal'
+
+// Response for fetching card transactions
+export interface ITransaction {
+  id: number
+  transactionId: string
+  ghResponseCode: string
+  cardScheme: number
+  type: number
+  createdAt: string
+  txStatus: string
+  vaultId: number
+  cardId: number
+  refTransactionId: string
+  responseCode: string | null
+  transactionAmount: string
+  transactionCurrency: string
+  billingAmount: string
+  billingCurrency: string
+  terminalId: string | null
+  wallet: number
+  transactionDateTime: string
+  processDateTime: string | null
+}
+
+export interface IPagination {
+  pageNumber: number
+  pageSize: number
+  totalPages: number
+  totalRecords: number
+}
+
+export interface IGetTransactionsResponse {
+  data: ITransaction[]
+  pagination: IPagination
+}
